requirements.txt
pandas==2.3.2
PyMySQL==1.1.2
python-dotenv==1.1.1
Scrapy==2.13.3
SQLAlchemy==2.0.43